add implementation for priority download queue
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@1432 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/rsiface/RemoteDirModel.cpp
@@ -765,7 +765,7 @@ void RemoteDirModel::downloadSelected(QModelIndexList list)
         /* if it is a dir, copy all files included*/
         else if (details.type == DIR_TYPE_DIR)
         {
-        	size_t prefixLen = details.path.rfind(details.name);
+        	int prefixLen = details.path.rfind(details.name);
         	if (prefixLen < 0) continue;
         	downloadDirectory(details, prefixLen);
         }
